Fixed compiling the code in Ubuntu 14.04
Apparently Ubuntu 14.04 still runs Qt 5.2 instead of Qt 5.4, so some parts of
the code had to be altered to make it be able to be compiled.

Strangely enough, it still doesn't work and I can't find the issue why. The
changes I made should not break anything for other systems, but now
track-o-bot can be compiled using qmake-qt4.

Also some functionality isn't present in Ubuntu. For instance notifications
don't work.